Added Messages component
