+++ lib/components/Application.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react'
-import firebase, { reference, signIn } from '../firebase';
+import firebase, { messsagesFromDatabase, signIn } from '../firebase';
 import { pick, map, extend } from 'lodash';
 import UserInput from './UserInput.js';
+import Messages from './Messages.js'
 
 // Very few things in this component are a good idea.
 // Feel free to blow it all away.
@@ -10,35 +11,29 @@ export default class Application extends Component {
   constructor() {
     super();
     this.state = {
-      messages: [],
       user: null
     }
   }
+
   componentDidMount() {
-    reference.limitToLast(100).on('value', (snapshot) => {
-      const messages = snapshot.val() || {};
-      this.setState({
-        messages: map(messages, (val, key) => extend(val, { key }))
-      });
-    });
     firebase.auth().onAuthStateChanged(user => this.setState({ user }));
   }
+
   addNewMessage(draftMessage) {
     const { user } = this.state;
-    reference.push({
+    messsagesFromDatabase.push({
       user: pick(user, 'displayName', 'email', 'uid'),
       content: draftMessage,
       createdAt: Date.now()
     });
   }
+
   render() {
     const { user, messages } = this.state;
     return (
       <div className="Application">
         {user ? <p>Hello {user.displayName}</p> : <button onClick={() => signIn()}>Sign In</button> }
-        <ul>
-          { this.state.messages.map(m => <li key={m.key}>{m.user.displayName}: {m.content}</li>) }
-        </ul>
+        <Messages />
         <UserInput addNewMessage={ this.addNewMessage.bind(this) }/>
       </div>
     )
