+++ lib/components/UserInput.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react'
-import firebase, { reference, signIn } from '../firebase';
+import firebase, { messsagesFromDatabase, signIn } from '../firebase';
 import { pick, map, extend } from 'lodash';
 
 export default class UserInput extends Component {
@@ -17,7 +17,6 @@ export default class UserInput extends Component {
     })
   }
   render() {
-    let characterCount = 140-this.state.draftMessage.length;
     return (
     <form className="message-input">
       <input
@@ -28,7 +27,7 @@ export default class UserInput extends Component {
         onChange={(e) => this.setState({ draftMessage: e.target.value })}
         />
       <input className='submit-button' type='submit' onClick={(e) => this.submitMessage(e)} value='Add New Message' />
-      <span className='character-count'>{characterCount}</span>
+      <span className='character-count'>{ 140-this.state.draftMessage.length}</span>
     </form>
     )
   }
