+++ lib/firebase.js
@@ -16,4 +16,4 @@ const provider = new firebase.auth.GoogleAuthProvider();
 
 export default firebase;
 export const signIn = () => auth.signInWithPopup(provider);
-export const reference = firebase.database().ref('messages');
+export const messsagesFromDatabase = firebase.database().ref('messages');
